Identity menu and content changed (improvements)
This commit involves rewrites and reorganization for affected pages, so that they are easier to consume and improve user experience.

Also, the navigation menu for Veracity docs has been changed. Now, the first item is a page "What is new" that gathers all the news and, at the bottom, has a place for a changelog. Note that this page has consumed "New Login URL", "API Security Flaw", and "Rotation of the Token Signing Key" because they are news.

Also, "Claims" are included in the "Overview" page now because that's something you would consult quite often, so it's worth bringing it up in the menu and adding to the overview page that offers the info you refer to the most.

This change also adds a "Client v1" page and a new page on the Single Sign On (that integrates about four separate pages we had before). Note that these pages used to be in the "Onboarding" section that I want to remove in #198 pull request.

The last change is creating a section for "Services API" and including "Policy Service" at the bottom because the Policy Service API is a part of the Services API, and it used to be apart (in the menu) from its parent.