Changes in the onboarding section + added image in dataworkbench section
For the Onboarding section, replaced the content of the onboarding.md and technicalrequirements.md with the content of the corresponding files from the get started section. This is temporary since the files in question were supposed to be replaced by the files in the get started section, but this section won't be published soon.

For data workbench, added an image to the authentication.md.