Update bookmarks.md
I've removed the mentions "If available", because the page says at the top that admin users can enable bookmarks, so if you don't see a bookmark icon, it's not enabled. However, we could add under the first paragraph something like "If you do not see a bookmark icon, ask your VAP administrator if bookmarks are enabled".